refactor: Implement base state for inheriting shared strategies across bot states
